requirements.txt
langchain>=0.1.16
langchain-openai>=0.1.3
openai>=1.21.0
python-dotenv>=1.0.1
requests>=2.32.3
pandas>=2.2.2
ta>=0.11.0
